modal in DBflowContainer -co-authored @rococoKid
