+++ frontend/react/containers/DBFlowContainer.jsx
@@ -1,4 +1,5 @@
 import React from "react";
+import { useState } from "react";
 import DBFlow from "../components/DBFlow.jsx";
 import ReactModal from 'react-modal';
 import DBQuery from "../components/DBQuery.jsx";
@@ -10,6 +11,13 @@ import { useConnectMutation } from '../../apiSlice.js';
 const DBFlowContainer = () => {
     // const count = useSelector((state) => state.database.value);
     // const dispatch = useDispatch();
+    
+    //useState function  for the modal being open:
+  const [modalIsOpen, setIsOpen] = useState(false);
+  // function to open the modal
+  const openModal = () => {
+      setIsOpen(true);
+    }
     
     // Using a query hook automatically fetches data and returns query values
   const [postConnect, { data, error, isLoading, isSuccess }] = useConnectMutation({
@@ -27,6 +35,8 @@ const DBFlowContainer = () => {
     return (     
         <div className="chart-page-container">
           <DBQuery />
+           {/* button to open modal: */}
+          <button className='button' onClick={openModal}>join</button>
           <ReactModal isOpen={modalIsOpen} shouldCloseOnEsc={true}>
             <div>this is our join modal</div>
           </ReactModal>
